feat: Run only in specified branches
By default, in main or master. Those branches are usually used for publishing versions.

BREAKING CHANGE: If you used this action for branches `main` or `master`, your configuration will continue working, because those branches are enabled by default. If you used it for other branches, specify all their names in the input `branches`. If you used this action on all branches, set the input `branches` to `*`.